+++ CHANGELOG.md
@@ -14,8 +14,9 @@ Improvements:
     * compiler: support for variadic function argument passing
     * compiler: NEW operator added
     * compiler: EXTENDS, BASE, IMPLEMENTS keywords added (OOP inheritance)
+    * compiler: support for virtual functions added (OOP)
     * compiler: PROPERTY support added (OOP)
-    * compiler: [breaking] CLASS/END CLASS, INTERFACE/END INTERFACE added (OOP virtual methods)
+    * compiler: [breaking] CLASS/END CLASS, INTERFACE/END INTERFACE added
     * compiler: generate proper error message when a constant is assigned to (#24)
 
 Bug Fixes: